pyproject.toml
[project]
name = "parakeetx-api-server"
version = "0.1.0"
description = "OpenAI-compatible FastAPI server for NVIDIA Parakeet ASR with optional pyannote diarization"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "ParakeetX" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.3.0",
  "numpy>=1.26.0",
  "soundfile>=0.12.1",
]

[project.optional-dependencies]
nemo = [
  "nemo-toolkit[asr]>=2.0.0",
]
diarization = [
  "pyannote.audio>=3.3.0",
]
cuda = [
  "torch>=2.3.0",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/parakeetx_api_server"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[project.scripts]
parakeetx-api-server = "parakeetx_api_server.main:app"
